+++ src/main/java/io/github/mac_genius/pgen/RunMode.java
@@ -1,0 +1,9 @@
+package io.github.mac_genius.pgen;
+
+/**
+ * Created by Mac on 10/13/16.
+ */
+public enum RunMode {
+    KATTIS,
+    LOCAL;
+}