+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # ProjectGenerator Changelog
 
+## Version 0.0.2 (10/14/2016)
+
+- [John Harrison] - Add support for C++ files
+- [John Harrison] - Add local and Kattis modes
+- [John Harrison] - Kattis mode downloads the example test cases
+- [John Harrison] - Kattis mode creates a shell script for easier testing of project
+
 ## Version 0.0.1 (10/12/2016)
 
 - [John Harrison] - One simple java template